Added page operation slice to reduce state functions being passed between components, fixed some banner locations + updated followers component

+++ localsound.backend/Infrastructure/Repositories/TrackRepository.cs
@@ -46,7 +46,12 @@ namespace localsound.backend.Infrastructure.Repositories
             try
             {
                 _dbContext.FileContent.Remove(track.TrackData);
-                _dbContext.FileContent.Remove(track.TrackImage);
+
+                if (track.TrackImage != null)
+                {
+                    _dbContext.FileContent.Remove(track.TrackImage);
+                }
+
                 _dbContext.ArtistTrackUpload.Remove(track);
 
                 await _dbContext.SaveChangesAsync();
